feat: dev 1차 merge (#8)
